# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sports": {
      "get": {
        "description": "Devuelve únicamente deportes activos.",
        "operationId": "CatalogController_sports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SportResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Listar deportes activos",
        "tags": [
          "Deportes"
        ]
      },
      "post": {
        "operationId": "CatalogController_createSport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Crear el deporte",
        "tags": [
          "Deportes"
        ]
      }
    },
    "/api/sports/{id}": {
      "get": {
        "description": "Lectura por UUID. La lectura individual no filtra por estado activo.",
        "operationId": "CatalogController_sport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Obtener el deporte",
        "tags": [
          "Deportes"
        ]
      },
      "patch": {
        "description": "Actualización parcial. Omitir campos que no cambian; no se aceptan null ni campos extra.",
        "operationId": "CatalogController_updateSport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Actualizar el deporte",
        "tags": [
          "Deportes"
        ]
      },
      "delete": {
        "description": "No elimina en cascada. Si tiene dependencias, desactivar el recurso o resolverlas antes.",
        "operationId": "CatalogController_deleteSport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado. Sin cuerpo de respuesta."
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Eliminar el deporte",
        "tags": [
          "Deportes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API del turnero de polideportivos",
    "description": "Flujo: deporte → contacto → zona y sede → detalle → fecha y turno → vista previa de WhatsApp. Los catálogos empiezan vacíos. Cargarlos con clave de gestión en el orden deportes, zonas, sedes, relaciones sede/deporte y turnos. Los UUID de los ejemplos son ficticios. Generar WhatsApp no envía mensajes, no bloquea disponibilidad y no confirma reservas. Los borradores se acceden mediante UUID secreto, sin login. JSON de hasta 16 KiB; 100 solicitudes/minuto/IP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ManagementKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clave de gestión solo para operaciones protegidas. No se almacena en la documentación."
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Código de estado HTTP."
          },
          "message": {
            "description": "Mensaje de negocio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "El turno no está disponible"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nombre del error; puede omitirse, por ejemplo en el límite de solicitudes."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "SportResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 120,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "Fútbol"
          },
          "icon": {
            "type": "string",
            "description": "Ícono del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "⚽"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el deporte está habilitado.",
            "example": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador público del deporte",
            "example": "00000000-0000-4000-8000-000000000001"
          }
        },
        "required": [
          "name",
          "icon",
          "isActive",
          "id"
        ]
      },
...
      "CreateSportDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 120,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "Fútbol"
          },
          "icon": {
            "type": "string",
            "description": "Ícono del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "⚽"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el deporte está habilitado.",
            "example": true
          }
        },
        "required": [
          "name",
          "icon",
          "isActive"
        ],
        "additionalProperties": false
      },
      "UpdateSportDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 120,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "Fútbol"
          },
          "icon": {
            "type": "string",
            "description": "Ícono del deporte. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "⚽"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el deporte está habilitado.",
            "example": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}